Use explicit price check for the V2 deposit instead of relying on the router
Previousluy the v2 deposit price limits check used amountAMin and amountBMin.
That has been changed to an explicit price check after the deposit.

The reason the previous v2 deposit price limits check is not good enough is
that it uses amountADesired/amountBDesired its reference price and not the
quote reference price as we would have wanted.

The code did attemp to make up for that by "shifting" amountAmin and
amountBmin to center around the quote reference, but that doesn't fully work
since the v2 router price limit check validates only one of the min or max
prices and not both.

The router chooses whether to validate the min or max
price according to the direction that the price moves vs the price of
amountADesired / amountBDesired (which is in our case the v3 price we
got on the first swap).
Can see that in the amountAMin and amountBMin parameters documentation here:
https://docs.uniswap.org/contracts/v2/reference/smart-contracts/router-02#addliquidity

Thus there would be no limit enforcemenet for some situations. For example:

[   <-quote->   ]   v3   v2
Where `quote` represent the quote reference price, the range around it is `zen`,
v3 is the price we got on univ3 and v2 is the v2 deposit price.

In such a case since the v2 price is out of the zen range the tx should fail,
but since the price went up in comparison to v3, the price check would only
be to the down side (and would therefore succeed).